langchain-course-project-agents-under-the-hood/langchain-course-project-agents-under-the-hood/pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.9",
    "langchain-tavily>=0.2.0",
    "langgraph>=1.0.0",
    "langsmith>=0.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "simpleeval>=0.9.13",
    "streamlit>=1.40.0",
]
